Move Recur to new property value type system
