+++ ical/todo.py
@@ -10,8 +10,7 @@ from pydantic import Field, root_validator
 from ._types import ComponentModel, Priority, RequestStatus, validate_until_dtstart
 from .alarm import Alarm
 from .parsing.property import ParsedProperty
-from .recur import Recur
-from .types import CalAddress, Geo, Uri
+from .types import CalAddress, Geo, Recur, Uri
 from .types.const import Classification, TodoStatus
 from .util import dtstamp_factory, normalize_datetime, uid_factory
